Draw the UI more efficiently
CopyResource is no longer needed! The problem was this sequence of
events:

1. OMSetRenderTargets(imgui) done during imgui render
2. PSSetShaderResources(imgui) in draw.rs
3. OMSetRenderTargets() in draw.rs

Step 2 fails (silently) because the imgui texture is already in use as a
render target, so it cannot also be a shader resource at the same time.
(This is checked immediately, not during the draw call.) The fix was to
move step 3 before step 2.